+++ cli/azd/pkg/ext/hooks_manager_test.go
@@ -3,6 +3,8 @@ package ext
 import (
 	"os"
 	"path/filepath"
+	"regexp"
+	"strings"
 	"testing"
 
 	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
@@ -103,11 +105,17 @@ func ensureScriptsExist(t *testing.T, configs map[string]*HookConfig) {
 	for _, hook := range configs {
 		ext := filepath.Ext(hook.Run)
 
-		if ext != "" {
+		if isValidFileExtension(ext) {
 			err := os.MkdirAll(filepath.Dir(hook.Run), osutil.PermissionDirectory)
 			require.NoError(t, err)
 			err = os.WriteFile(hook.Run, nil, osutil.PermissionExecutableFile)
 			require.NoError(t, err)
 		}
 	}
+}
+
+var fileExtensionRegex = regexp.MustCompile(`^\.[\w]{1,4}$`)
+
+func isValidFileExtension(extension string) bool {
+	return strings.ToLower(extension) != "" && fileExtensionRegex.MatchString(extension)
 }