+++ cli/azd/pkg/infra/provisioning/bicep/bicep_provider.go
@@ -98,8 +98,16 @@ func (p *BicepProvider) Initialize(ctx context.Context, projectPath string, opti
 }
 
 func (p *BicepProvider) State(ctx context.Context) (*StateResult, error) {
+	var err error
+	spinnerMessage := "Loading Bicep template"
 	// TODO: Report progress, "Loading Bicep template"
-	p.console.ShowSpinner(ctx, "Loading Bicep template", input.Step)
+	p.console.ShowSpinner(ctx, spinnerMessage, input.Step)
+	defer func() {
+		// Make sure we stop the spinner if an error occurs with the last message.
+		if err != nil {
+			p.console.StopSpinner(ctx, spinnerMessage, input.StepFailed)
+		}
+	}()
 
 	modulePath := p.modulePath()
 	_, template, err := p.compileBicep(ctx, modulePath)
@@ -113,7 +121,8 @@ func (p *BicepProvider) State(ctx context.Context) (*StateResult, error) {
 	}
 
 	// TODO: Report progress, "Retrieving Azure deployment"
-	p.console.ShowSpinner(ctx, "Retrieving Azure deployment", input.Step)
+	spinnerMessage = "Retrieving Azure deployment"
+	p.console.ShowSpinner(ctx, spinnerMessage, input.Step)
 
 	armDeployment, err := latestCompletedDeployment(ctx, p.env.GetEnvName(), scope)
 	if err != nil {
@@ -130,7 +139,8 @@ func (p *BicepProvider) State(ctx context.Context) (*StateResult, error) {
 	}
 
 	// TODO: Report progress, "Normalizing output parameters"
-	p.console.ShowSpinner(ctx, "Normalizing output parameters", input.Step)
+	spinnerMessage = "Normalizing output parameters"
+	p.console.ShowSpinner(ctx, spinnerMessage, input.Step)
 
 	state.Outputs = p.createOutputParameters(
 		template.Outputs,
@@ -389,6 +399,12 @@ func (p *BicepProvider) Destroy(ctx context.Context, options DestroyOptions) (*D
 		return nil, fmt.Errorf("getting key vaults to purge: %w", err)
 	}
 
+	// TODO: Report progress, "Getting Managed HSMs to purge"
+	managedHSMs, err := p.getManagedHSMsToPurge(ctx, groupedResources)
+	if err != nil {
+		return nil, fmt.Errorf("getting managed hsms to purge: %w", err)
+	}
+
 	// TODO: Report progress, "Getting App Configurations to purge"
 	appConfigs, err := p.getAppConfigsToPurge(ctx, groupedResources)
 	if err != nil {
@@ -418,6 +434,13 @@ func (p *BicepProvider) Destroy(ctx context.Context, options DestroyOptions) (*D
 			return p.purgeKeyVaults(ctx, keyVaults, options, skipPurge)
 		},
 	}
+	managedHSMsPurge := itemToPurge{
+		resourceType: "Managed HSM",
+		count:        len(managedHSMs),
+		purge: func(skipPurge bool, self *itemToPurge) error {
+			return p.purgeManagedHSMs(ctx, managedHSMs, options, skipPurge)
+		},
+	}
 	appConfigsPurge := itemToPurge{
 		resourceType: "App Configuration",
 		count:        len(appConfigs),
@@ -434,7 +457,7 @@ func (p *BicepProvider) Destroy(ctx context.Context, options DestroyOptions) (*D
 	}
 
 	var purgeItem []itemToPurge
-	for _, item := range []itemToPurge{keyVaultsPurge, appConfigsPurge, aPIManagement} {
+	for _, item := range []itemToPurge{keyVaultsPurge, managedHSMsPurge, appConfigsPurge, aPIManagement} {
 		if item.count > 0 {
 			purgeItem = append(purgeItem, item)
 		}
@@ -778,6 +801,46 @@ func (p *BicepProvider) getKeyVaultsToPurge(
 	return vaultsToPurge, nil
 }
 
+func (p *BicepProvider) getManagedHSMs(
+	ctx context.Context,
+	groupedResources map[string][]azcli.AzCliResource,
+) ([]*azcli.AzCliManagedHSM, error) {
+	managedHSMs := []*azcli.AzCliManagedHSM{}
+
+	for resourceGroup, groupResources := range groupedResources {
+		for _, resource := range groupResources {
+			if resource.Type == string(infra.AzureResourceTypeManagedHSM) {
+				managedHSM, err := p.azCli.GetManagedHSM(ctx, azure.SubscriptionFromRID(resource.Id), resourceGroup, resource.Name)
+				if err != nil {
+					return nil, fmt.Errorf("listing managed hsm %s properties: %w", resource.Name, err)
+				}
+				managedHSMs = append(managedHSMs, managedHSM)
+			}
+		}
+	}
+
+	return managedHSMs, nil
+}
+
+func (p *BicepProvider) getManagedHSMsToPurge(
+	ctx context.Context,
+	groupedResources map[string][]azcli.AzCliResource,
+) ([]*azcli.AzCliManagedHSM, error) {
+	managedHSMs, err := p.getManagedHSMs(ctx, groupedResources)
+	if err != nil {
+		return nil, err
+	}
+
+	managedHSMsToPurge := []*azcli.AzCliManagedHSM{}
+	for _, v := range managedHSMs {
+		if v.Properties.EnableSoftDelete && !v.Properties.EnablePurgeProtection {
+			managedHSMsToPurge = append(managedHSMsToPurge, v)
+		}
+	}
+
+	return managedHSMsToPurge, nil
+}
+
 func (p *BicepProvider) getCognitiveAccountsToPurge(
 	ctx context.Context,
 	groupedResources map[string][]azcli.AzCliResource,
@@ -826,12 +889,30 @@ func (p *BicepProvider) purgeKeyVaults(
 	skip bool,
 ) error {
 	for _, keyVault := range keyVaults {
-		err := p.runPurgeAsStep(ctx, "key vault", keyVault.Name, func() error {
+		err := p.runPurgeAsStep(ctx, "Key Vault", keyVault.Name, func() error {
 			return p.azCli.PurgeKeyVault(
 				ctx, azure.SubscriptionFromRID(keyVault.Id), keyVault.Name, keyVault.Location)
 		}, skip)
 		if err != nil {
 			return fmt.Errorf("purging key vault %s: %w", keyVault.Name, err)
+		}
+	}
+	return nil
+}
+
+func (p *BicepProvider) purgeManagedHSMs(
+	ctx context.Context,
+	managedHSMs []*azcli.AzCliManagedHSM,
+	options DestroyOptions,
+	skip bool,
+) error {
+	for _, managedHSM := range managedHSMs {
+		err := p.runPurgeAsStep(ctx, "Managed HSM", managedHSM.Name, func() error {
+			return p.azCli.PurgeManagedHSM(
+				ctx, azure.SubscriptionFromRID(managedHSM.Id), managedHSM.Name, managedHSM.Location)
+		}, skip)
+		if err != nil {
+			return fmt.Errorf("purging managed hsm %s: %w", managedHSM.Name, err)
 		}
 	}
 	return nil